platforms: Fixed the type of `app_address`, fixing a clang-tidy lint item

+++ src/platforms/swlink/usbdfu.c
@@ -26,7 +26,7 @@
 #include "usbdfu.h"
 #include "platform.h"
 
-uint32_t app_address = 0x08002000U;
+uintptr_t app_address = 0x08002000U;
 uint32_t rev;
 
 void dfu_detach(void)
